requirements.txt
fastapi
uvicorn
pandas
openpyxl
psycopg[binary]
python-multipart
pydantic
python-dotenv
PyJWT
passlib[bcrypt]